Renamed paths/files from binary -> beve
