+++ include/glaze/beve.hpp
@@ -1,0 +1,11 @@
+// Glaze Library
+// For the license information refer to glaze.hpp
+
+#pragma once
+
+#include "glaze/beve/custom.hpp"
+#include "glaze/beve/header.hpp"
+#include "glaze/beve/ptr.hpp"
+#include "glaze/beve/read.hpp"
+#include "glaze/beve/wrappers.hpp"
+#include "glaze/beve/write.hpp"
