+++ include/glaze/beve/beve_to_json.hpp
@@ -3,7 +3,7 @@
 
 #pragma once
 
-#include "glaze/binary/header.hpp"
+#include "glaze/beve/header.hpp"
 #include "glaze/json/write.hpp"
 
 namespace glz
