--- include/glaze/binary.hpp
@@ -1,11 +1,0 @@
-// Glaze Library
-// For the license information refer to glaze.hpp
-
-#pragma once
-
-#include "glaze/binary/custom.hpp"
-#include "glaze/binary/header.hpp"
-#include "glaze/binary/ptr.hpp"
-#include "glaze/binary/read.hpp"
-#include "glaze/binary/wrappers.hpp"
-#include "glaze/binary/write.hpp"
